Update: add comments and move io functions to its coresponding file
